Update library.json
First of a few PRs to start adding styles we archived to test the docs a while back, plus improvements we added, and new styles. We need to improve our own release steps for these, and I have loads more that I'm working through in alphabetical order, just wanted to do one first to see if the style was correct.